[WEB-440] feat: create project feature selection modal. * [WEB-399] chore: explain project identifier.
